fix iOS sensorSettings and some refactors
iOS plugin class was set with the wrong call to sensorSettings, this configuration wasn't taking effect. Also, some refactors in the definitions of the subclasses were done (sensorOrientation, sensorLuminosity and sensorStability), since the message parameter is now being defined on MessageSettings.
On the Android plugin class: code style and prevention of null pointer exception on CaptureStages definition